+++ src/components/form/fields/schema-field.ts
@@ -7,6 +7,7 @@ import { resetDependentFields } from './field-helpers';
 import { FieldTemplate } from '../templates';
 import { getHelpComponent } from '../help';
 import { FormSchema } from '../form.types';
+import { TimePicker } from '../widgets/time-picker';
 
 /**
  * If given a value and schema, check if the value should be translated
@@ -278,6 +279,13 @@ export class SchemaField extends React.Component<FieldProps> {
             ...this.props,
             onChange: this.handleChange,
         };
+
+        if (this.props.schema.format === 'time') {
+            fieldProps.uiSchema = {
+                'ui:widget': TimePicker,
+                ...fieldProps.uiSchema,
+            };
+        }
 
         return React.createElement(JSONSchemaField, fieldProps);
     }